[UPDATE] Linux compile fix for literals.

+++ engine/utility/public/bytes.hpp
@@ -2,17 +2,17 @@
 
 #include <cstdint>
 
-constexpr uint64_t operator"" _kb(uint64_t kilobytes)
+constexpr uint64_t operator""_kb(unsigned long long kilobytes)
 {
     return kilobytes * 1024;
 }
 
-constexpr uint64_t operator"" _mb(uint64_t megabytes)
+constexpr uint64_t operator""_mb(unsigned long long megabytes)
 {
     return megabytes * 1024 * 1024;
 }
 
-constexpr uint64_t operator"" _gb(uint64_t gigabytes)
+constexpr uint64_t operator""_gb(unsigned long long gigabytes)
 {
     return gigabytes * 1024 * 1024 * 1024;
 }
